Move opengarage coordinator to separate module
