+++ homeassistant/components/opengarage/__init__.py
@@ -2,22 +2,15 @@
 
 from __future__ import annotations
 
-from datetime import timedelta
-import logging
-from typing import Any
-
 import opengarage
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST, CONF_PORT, CONF_VERIFY_SSL, Platform
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import update_coordinator
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from .const import CONF_DEVICE_KEY, DOMAIN
-
-_LOGGER = logging.getLogger(__name__)
+from .coordinator import OpenGarageDataUpdateCoordinator
 
 PLATFORMS = [Platform.BINARY_SENSOR, Platform.BUTTON, Platform.COVER, Platform.SENSOR]
 
@@ -49,32 +42,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data[DOMAIN].pop(entry.entry_id)
 
     return unload_ok
-
-
-class OpenGarageDataUpdateCoordinator(DataUpdateCoordinator[dict[str, Any]]):  # pylint: disable=hass-enforce-coordinator-module
-    """Class to manage fetching Opengarage data."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        *,
-        open_garage_connection: opengarage.OpenGarage,
-    ) -> None:
-        """Initialize global Opengarage data updater."""
-        self.open_garage_connection = open_garage_connection
-
-        super().__init__(
-            hass,
-            _LOGGER,
-            name=DOMAIN,
-            update_interval=timedelta(seconds=5),
-        )
-
-    async def _async_update_data(self) -> dict[str, Any]:
-        """Fetch data."""
-        data = await self.open_garage_connection.update_state()
-        if data is None:
-            raise update_coordinator.UpdateFailed(
-                "Unable to connect to OpenGarage device"
-            )
-        return data
